fix readthedocs, restructure pkg (#3)
* fix readthedocs, restructure pkg

* Update src/simplay/src/simplay/core.py
