+++ .gitignore
@@ -1,1 +1,3 @@
 **dist
+**build
+**__pycache__
