+++ docs/source/api_reference/simplay.core.rst
@@ -1,0 +1,15 @@
+==============================================
+``simplay.core`` -- Core components of Simplay
+==============================================
+
+.. automodule:: simplay.core
+   
+   .. autoclass:: VisualEnvironment
+       :members:
+       :show-inheritance:
+    
+   .. autoclass:: VisualComponent
+       :members:
+
+   .. autoclass:: VisualizationManager
+       :members:
