+++ docs/source/api_reference/simplay.events.rst
@@ -1,6 +1,6 @@
-=====================================
+=========================================
 ``simplay.events`` --- Visual Event Types
-=====================================
+=========================================
 
 .. automodule:: simplay.events
