+++ docs/source/api_reference/simplay.rst
@@ -1,6 +1,6 @@
-=========
+===========
 ``simplay``
-=========
+===========
 
 .. automodule:: simplay
