+++ docs/source/api_reference/simplay.visualization.rst
@@ -3,9 +3,6 @@
 ===========================================
 
 .. automodule:: simplay.visualization
-
-    .. autoclass:: VisualizationManager
-       :inherited-members:
     
     .. autoclass:: VisualGrid
        :inherited-members:
