+++ docs/source/api_reference/simplay.visualutil.rst
@@ -5,9 +5,13 @@
 .. automodule:: simplay.visualutil
    
     .. autoclass:: BasicVisualUtil
+       :members:
 
     .. autoclass:: ResourceVisualUtil
+       :members:
     
     .. autoclass:: ContainerVisualUtil
+       :members:
     
     .. autoclass:: StoreVisualUtil
+       :members:
