+++ docs/source/contents.rst
@@ -1,0 +1,23 @@
+.. _contents:
+
+=========================
+Documentation for SimPlay
+=========================
+
+.. only:: html
+
+    Contents:
+
+.. toctree::
+   :maxdepth: 2
+
+   index
+   usage
+   api_reference/index
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`search`
