+++ docs/source/usage.rst
@@ -8,16 +8,16 @@ Installation
 
 To use Simplay, first decide how you want to use it.
 Three components exist:
- - SimPlay (the python library allowing you to generate the EventQueue)
- - SimPlay-Web (the npm packages allowing you to render the EventQueue)
- - SimPlay-Jupyter (the python package allowing you to use both within jupyter)
+- SimPlay (the python library allowing you to generate the SimulationData)
+- SimPlay-Web (the npm packages allowing you to render the SimulationData)
+- SimPlay-Jupyter (the python package allowing you to use both within jupyter notebooks)
 
 
 To install SimPlay:
 
 .. code-block:: console
 
-   (.venv) $ pip install SimPlay
+   (.venv) $ pip install simplay
 
 
 To install SimPlay-Web:
@@ -30,4 +30,4 @@ To install the jupyter extension:
 
 .. code-block:: console
 
-   $ pip install SimPlay-Jupyter
+   $ pip install simplay-jupyter
